Remove references to libqwtaddons - we are not using that anymore
libqwtaddons was a library to provide shims for features in QWT that
were templated. PySide2 presently doesn't support templates, so this
shim library was created to bridge that gap. With this release version
of pysideqwt, we didn't think it was good to keep the shim library, as
it was very project-specific.

Instead if features that require templated parts of QWT are needed, we
should add templating support to Shiboken - at least in the form of
static instantiation of template classes in Python.